bank processing selection working

+++ main.py
@@ -6,18 +6,36 @@ import glob
 import os
 
 # TODO:
-# - Parse Data and organize by month
+# - Automate getting monthly data
 
-# acc = gspread.service_account()
+acc = gspread.service_account()
 
-# sheet = acc.open("personal-finances")
-# pendingFiles = glob.glob("./finance/*.csv")
+sheet = acc.open("personal-finances")
+pendingFiles = glob.glob("./finance/*.csv")
 
 # Testing
 # sheet = acc.open("personal-finances-test")
-pendingFiles = glob.glob("./test-data/*.csv")
+# pendingFiles = glob.glob("./test-data/*.csv")
 
 months = ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"]
+
+def turnToList(strings):
+    arr = []
+    if strings == "" or strings == "all":
+        arr = list(["apple", "discover", "bofa"])
+    elif re.search(".*,\d*", strings):
+        strings = strings.replace(" ", "")
+        arr = list(strings.split(","))
+    elif re.search("[ ]", strings):
+        arr = list(strings.split(" "))
+    elif re.search("-", strings):
+        arr = list(strings.split("-"))
+    elif len(strings) > 0:
+        arr = list()
+        arr.append(strings)
+    else:
+        print(strings + " was not converted to list")
+    defineProcess(arr)
 
 
 def addTotal(data):
@@ -85,7 +103,7 @@ def loadFile(file, cfg):
         next(csv_file)
         csv_reader = csv.reader(csv_file)
         for row in csv_reader:
-            # wksName = sheet.worksheet(re.search("\d{4}", row[0]).group(0))
+            wksName = sheet.worksheet(re.search("\d{4}", row[0]).group(0))
             date = row[cfg["date"]]
             txtDate = textDate(row[cfg["date"]])
             name = row[cfg["name"]]
@@ -102,61 +120,31 @@ def loadFile(file, cfg):
 
 def loadCSV(rows, cfg={"txtDate": 0, "date": 1, "name": 2, "amount": 3, "category": 5, "rAmount": 4, "cfg": 6}):
     for row in rows:
-        print([row[cfg["txtDate"]], row[cfg["date"]], row[cfg["name"]], row[cfg["amount"]],row[cfg["category"]], row[cfg["rAmount"]], row[cfg["cfg"]]], 2)
-        # row[7].insert_row([row[cfg["txtDate"]], row[cfg["date"]], row[cfg["name"]], row[cfg["amount"]],row[cfg["category"]], row[cfg["rAmount"]], row[cfg["cfg"]]], 2)
+        # print([row[cfg["txtDate"]], row[cfg["date"]], row[cfg["name"]], row[cfg["amount"]],row[cfg["category"]], row[cfg["rAmount"]], row[cfg["cfg"]]], 2)
+        row[7].insert_row([row[cfg["txtDate"]], row[cfg["date"]], row[cfg["name"]], row[cfg["amount"]],row[cfg["category"]], row[cfg["rAmount"]], row[cfg["cfg"]]], 2)
         time.sleep(2)
 
 def defineProcess(banks):
-    # discover = any
-    # bofa = any
-    # apple = any
     for file in pendingFiles:
-        match banks:
-            case "Discover":
-                discover = re.search(".*Discover(.+).csv", file)
-            case "apple":
-                apple = re.search(".*Apple(.+).csv", file)
-            case "bofa":
+        for bank in banks:
+            if bank == "bofa":
                 bofa = re.search("(.+)_4764.csv", file)
-            case default:
-                print("invalid bank")
-        # bofa = re.search("(.+)_4764.csv", file)
-        # apple = re.search(".*Apple(.+).csv", file)
-        # discover = re.search(".*Discover(.+).csv", file)
+                if bofa:
+                    rows = loadFile(file, {"category": 1, "amount": 4, "name": 2, "date": 0, "cfg": "bofa"})
+                    loadCSV(rows)
+                    os.remove(file)
+            if bank == "apple":
+                apple = re.search(".*Apple(.+).csv", file)
+                if apple:
+                    rows = loadFile(file, {"category": 4, "amount": 6, "name": 2, "date": 0, "cfg": "apple"})
+                    loadCSV(rows)
+                    os.remove(file)
+            if bank == "discover":
+                discover = re.search(".*Discover(.+).csv", file)
+                if discover:
+                    rows = loadFile(file, {"category": 4, "amount": 3, "name": 2, "date": 0, "cfg": "discover"})
+                    loadCSV(rows, {"txtDate": 0, "date": 1, "name": 2, "amount": 3, "rAmount": 4, "category": 5, "cfg": 6})
+                    os.remove(file)
 
-
-        if discover:
-            rows = loadFile(file, {"category": 4, "amount": 3, "name": 2, "date": 0, "cfg": "discover"})
-            loadCSV(rows, {"txtDate": 0, "date": 1, "name": 2, "amount": 3, "rAmount": 4, "category": 5, "cfg": 6})
-            os.remove(file)
-        elif apple:
-            rows = loadFile(file, {"category": 4, "amount": 6, "name": 2, "date": 0, "cfg": "apple"})
-            loadCSV(rows)
-            os.remove(file)
-        elif bofa:
-            rows = loadFile(file, {"category": 1, "amount": 4, "name": 2, "date": 0, "cfg": "bofa"})
-            loadCSV(rows)
-            os.remove(file)
-        else:
-            # time.sleep(1)
-            os.remove(file)
-
-def turnToList(strings):
-    arr = []
-
-    if re.search(".*,\d*", strings):
-        strings = strings.replace(" ", "")
-        arr = list(strings.split(","))
-        print(arr)
-    elif re.search("[ ]", strings):
-        arr = list(strings.split(" "))
-        print(arr)
-    elif re.search("-", strings):
-        arr = list(strings.split("-"))
-        print(arr)
-    else:
-        print(strings + " not detected")
-
-user = input("Enter Config: ")
-defineProcess(user)
+turnToList(input("Enter Config: "))
